fix(tests): Fix type conversion warnings in JSON tests
Fixed compilation errors with -Werror:
- Use dap_json_object_new_int64() for INT64_MAX instead of new_int()
- Cast int constants to size_t when comparing with size_t variables

Files fixed:
- test_boundary_conditions.c: INT64_MAX overflow fix
- test_memory_dos.c: size_t comparison fixes (2 places)
- test_array_edge_cases.c: size_t comparison fix
- test_serialization_round_trip.c: size_t comparison fix

All tests now compile without warnings under -Werror.

+++ tests/unit/json/test_array_edge_cases.c
@@ -260,7 +260,7 @@ static bool s_test_large_arrays(void) {
     DAP_TEST_FAIL_IF_NULL(l_json, "Parse large array");
     
     size_t len = dap_json_array_length(l_json);
-    DAP_TEST_FAIL_IF(len != ELEMENT_COUNT, "Large array has correct length");
+    DAP_TEST_FAIL_IF(len != (size_t)ELEMENT_COUNT, "Large array has correct length");
     
     // Spot check elements
     int first = dap_json_array_get_int(l_json, 0);
